test(refactor): extract domain factories from conftest to tests/factories/
New modules:
- tests/factories/adapters.py
- tests/factories/agents.py
- tests/factories/bootstrap.py
- tests/factories/messages.py
- tests/factories/nats_server.py
- tests/factories/plugins.py
- tests/factories/pools.py
- tests/factories/stores.py

Conftest files now re-export from factories for backward compatibility.
Removed empty tests/cli/conftest.py and tests/llm/conftest.py.
