--- tests/cli/conftest.py
@@ -1,1 +1,0 @@
-"""Shared fixtures and helpers for CLI tests."""
