+++ tests/factories/__init__.py
@@ -1,0 +1,11 @@
+"""Domain-specific test factories for the Lyra test suite."""
+
+# Re-exports for backward compatibility
+from tests.factories.adapters import *  # noqa: F403
+from tests.factories.agents import *  # noqa: F403
+from tests.factories.bootstrap import *  # noqa: F403
+from tests.factories.messages import *  # noqa: F403
+from tests.factories.nats_server import *  # noqa: F403
+from tests.factories.plugins import *  # noqa: F403
+from tests.factories.pools import *  # noqa: F403
+from tests.factories.stores import *  # noqa: F403
